+++ packages/project/src/Workflow.ts
@@ -30,7 +30,16 @@ class Workflow {
     // history needs to be on workflow object.
     this.workflow.history = workflow.history?.length ? workflow.history : [];
 
-    const { id, name, openfn, steps, history, ...options } = workflow;
+    const {
+      id,
+      name,
+      openfn,
+      steps,
+      history,
+      start: _start,
+      options,
+      ...rest
+    } = workflow;
     if (!(id || name)) {
       throw new Error('A Workflow MUST have a name or id');
     }
@@ -45,13 +54,21 @@ class Workflow {
     }
 
     this.openfn = openfn;
-    this.options = options;
+    this.options = Object.assign({}, options, rest);
 
     this._buildIndex();
   }
 
   get steps(): WithMeta<l.Job & l.Trigger>[] {
     return this.workflow.steps;
+  }
+
+  get start(): string | undefined {
+    return this.workflow.start;
+  }
+
+  set start(s: string) {
+    this.workflow.start = s;
   }
 
   _buildIndex() {