+++ packages/project/test/parse/from-fs.test.ts
@@ -40,6 +40,7 @@ test.serial('should load workspace config from json', async (t) => {
 
   t.deepEqual(project.config, {
     x: 1,
+    credentials: 'credentials.yaml',
     dirs: { projects: '.projects', workflows: 'workflows' },
     formats: { openfn: 'json', project: 'json', workflow: 'json' },
   });
@@ -62,13 +63,14 @@ test.serial('should load workspace config from yaml', async (t) => {
   const project = await parseProject({ root: '/ws' });
 
   t.deepEqual(project.config, {
+    credentials: 'credentials.yaml',
     x: 1,
     dirs: { projects: '.projects', workflows: 'workflows' },
     formats: { openfn: 'yaml', project: 'yaml', workflow: 'yaml' },
   });
 });
 
-test.serial('should load single workflow', async (t) => {
+test.serial('should load single workflow in new flat format', async (t) => {
   mockFile('/ws/openfn.yaml', buildConfig());
 
   mockFile('/ws/workflows/my-workflow/my-workflow.yaml', {
@@ -80,6 +82,7 @@ test.serial('should load single workflow', async (t) => {
         expression: 'job.js',
       },
     ],
+    start: 'a',
   });
 
   mockFile('/ws/workflows/my-workflow/job.js', `fn(s => s)`);
@@ -92,7 +95,45 @@ test.serial('should load single workflow', async (t) => {
   t.truthy(wf);
   t.is(wf.id, 'my-workflow');
   t.is(wf.name, 'My Workflow');
+  t.is(wf.start, 'a');
 });
+
+// hmm, maybe I shouldn't support this, because it puts some wierd stuff in the code
+// and new CLI will just use the new  format
+test.serial(
+  'should load single workflow in old { workflow, options } format',
+  async (t) => {
+    mockFile('/ws/openfn.yaml', buildConfig());
+
+    mockFile('/ws/workflows/my-workflow/my-workflow.yaml', {
+      workflow: {
+        id: 'my-workflow',
+        name: 'My Workflow',
+        steps: [
+          {
+            id: 'a',
+            expression: 'job.js',
+          },
+        ],
+      },
+      options: {
+        start: 'a',
+      },
+    });
+
+    mockFile('/ws/workflows/my-workflow/job.js', `fn(s => s)`);
+
+    const project = await parseProject({ root: '/ws' });
+
+    t.is(project.workflows.length, 1);
+
+    const wf = project.getWorkflow('my-workflow');
+    t.truthy(wf);
+    t.is(wf.id, 'my-workflow');
+    t.is(wf.name, 'My Workflow');
+    t.is(wf.start, 'a');
+  }
+);
 
 test.serial('should load single workflow from json', async (t) => {
   mockFile(