+++ packages/engine-multi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # engine-multi
 
+## 1.10.1
+
+### Patch Changes
+
+- Updated dependencies [b262d10]
+- Updated dependencies [147a431]
+  - @openfn/runtime@1.8.1
+
 ## 1.10.0
 
 ### Minor Changes